hfti {lsei}	R Documentation

Least Squares Solution using Householder Transformation

Description

Solves the least squares problem using Householder forward triangulation with column interchanges. It is an R interface to the HFTI function that is described in Lawson and Hanson (1974, 1995). Its Fortran implementation is public domain and is available at http://www.netlib.org/lawson-hanson.

Usage

hfti(a, b, tol=1e-7) 

Arguments

a
Design matrix.

b
Response vector or matrix.

tol
Tolerance for determining the pseudorank.

Details

Given matrix a and vector b, hfti solves the least squares problem:

minimize || a x - b ||.

Value

b
first krank elements contains the solution

krank
psuedo-rank

rnorm
Euclidean norm of the residual vector.

Author(s)

Yong Wang <yongwang@auckland.ac.nz>

References

Lawson and Hanson (1974, 1995). Solving least squares problems. Englewood Cliffs, N.J., Prentice-Hall.

See Also

lsei, nnls.

Examples

a = matrix(rnorm(10*4), nrow=10)
b = a %*% c(0,1,-1,1) + rnorm(10)
hfti(a, b)

[Package lsei version 1.2-0.1 Index]